refactor(tools): move parent tool components into subfolders
