+++ src/components/tools/paint/PaintTool.vue
@@ -15,7 +15,7 @@ import { computed, defineComponent, PropType } from 'vue';
 import vtkWidgetManager from '@kitware/vtk.js/Widgets/Core/WidgetManager';
 import { usePaintToolStore } from '@/src/store/tools/paint';
 import { LPSAxisDir } from '@/src/types/lps';
-import PaintWidget2D from './paint/PaintWidget2D.vue';
+import PaintWidget2D from './PaintWidget2D.vue';
 
 export default defineComponent({
   name: 'PaintTool',
