+++ src/components/tools/rectangle/RectangleTool.vue
