+++ src/components/tools/ruler/RulerTool.vue
